+++ backend/api/src/unsubscribe.ts
@@ -4,7 +4,7 @@ import { getPrivateUser } from 'shared/utils'
 import { PrivateUser } from 'common/user'
 import { NOTIFICATION_DESCRIPTIONS } from 'common/notification'
 import { notification_preference } from 'common/user-notification-preferences'
-import { getApiUrl } from 'common/api'
+import { getApiUrl } from 'common//api/utils'
 import { trackPublicEvent } from 'shared/analytics'
 
 export const unsubscribe = async (req: Request, res: Response) => {
