+++ backend/discord-bot/src/api.ts
@@ -1,4 +1,4 @@
-import { FullMarket } from 'common/api-market-types'
+import { FullMarket } from 'common/api/market-types'
 import { ContractMetrics } from 'common/calculate-metrics'
 import { config } from 'discord-bot/constants/config'
 import { User } from 'discord.js'
