+++ backend/shared/src/complete-quest-internal.ts
@@ -8,7 +8,7 @@ import { runTxnFromBank } from 'shared/txn/run-txn'
 import { createSupabaseClient } from 'shared/supabase/init'
 import { getRecentContractIds } from 'common/supabase/contracts'
 import { getUserShareEventsCount } from 'common/supabase/user-events'
-import { APIError } from 'common/api'
+import { APIError } from 'common//api/utils'
 import { createQuestPayoutNotification } from 'shared/create-notification'
 import * as dayjs from 'dayjs'
 import * as utc from 'dayjs/plugin/utc'
