+++ web/components/chat-input.tsx
@@ -6,7 +6,7 @@ import { MAX_COMMENT_LENGTH } from 'common/comment'
 import { safeLocalStorage } from 'web/lib/util/local'
 import { useUser } from 'web/hooks/use-user'
 import { call } from 'web/lib/firebase/api'
-import { getApiUrl } from 'common/api'
+import { getApiUrl } from 'common/api/utils'
 import { Col } from 'web/components/layout/col'
 import { Extension } from '@tiptap/core'
 import { useEvent } from 'web/hooks/use-event'
