+++ web/components/comments/edit-comment-modal.tsx
@@ -8,7 +8,7 @@ import { Modal } from 'web/components/layout/modal'
 import { useEffect, useState } from 'react'
 import { User } from 'common/user'
 import { safeLocalStorage } from 'web/lib/util/local'
-import { getApiUrl } from 'common/api'
+import { getApiUrl } from 'common/api/utils'
 import { call } from 'web/lib/firebase/api'
 import { Title } from 'web/components/widgets/title'
 import { JSONContent } from '@tiptap/core'
