+++ web/pages/api/v0/search-markets.ts
@@ -1,4 +1,4 @@
-import { LiteMarket, toLiteMarket } from 'common/api-market-types'
+import { LiteMarket, toLiteMarket } from 'common/api/market-types'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { applyCorsHeaders } from 'web/lib/api/cors'
 import { APIError, searchContracts } from 'web/lib/firebase/api'
